+++ src/frontend/base.ts
@@ -15,28 +15,29 @@ export class StaticSeekError extends Error {}
 
 // index
 export type IndexOpt = {
-    key_fields?: string[];
-    search_targets?: string[];
+    key_fields?: Path[];
+    search_targets?: Path[];
     distance?: number;
+    weights?: [Path, number][];
 };
 
 export const SearchEnv_v = v.object({
     search_targets: v.optional(v.array(v.string())),
     field_names: v.record(v.string(), v.string()),
-    weight: v.number(),
+    weights: v.array(v.tuple([v.string(), v.number()])),
     distance: v.number(),
 });
 
 export type SearchEnv = v.InferOutput<typeof SearchEnv_v>;
 
-export const StaticIndex_v = v.object({
+export const StaticSeekIndexRoot_v = v.object({
     version: v.string(),
     type: v.string(),
     env: SearchEnv_v,
     index_entry: v.unknown(),
 });
 
-export type StaticIndex<T> = {
+export type StaticSeekIndexRoot<T> = {
     version: string;
     type: string;
     env: SearchEnv;
@@ -44,7 +45,6 @@ export type StaticIndex<T> = {
 };
 
 // query
-
 export type Reference = {
     token: string;
     path: Path;
@@ -69,7 +69,7 @@ export interface SearchIndex<T> {
     search(env: SearchEnv, keyword: string[]): Promise<SearchResult[]>;
 }
 
-export type StaticSeekIndex = StaticIndex<SearchIndex<unknown>>;
-export type StaticSeekIndexObject = StaticIndex<unknown>;
+export type StaticSeekIndex = StaticSeekIndexRoot<SearchIndex<unknown>>;
+export type StaticSeekIndexObject = StaticSeekIndexRoot<unknown>;
 export type SearchFnResult = Promise<SearchResult[] | StaticSeekError>;
 export type SearchFn = (query: string) => SearchFnResult;