+++ src/main.ts
@@ -1,7 +1,6 @@
 export { createIndex, indexToObject, createIndexFromObject } from "@src/frontend/indexing";
 export { search } from "@src/frontend/search";
 export { createSearchFn } from "@src/frontend/util";
-export { HybridBigramInvertedIndex } from "@src/method/hybridbigraminvertedindex";
 export { HybridTrieBigramInvertedIndex } from "@src/method/hybridtriebigraminvertedindex";
 export { LinearIndex } from "@src/method/linearindex";
 export { GPULinearIndex } from "@src/method/gpulinearindex";
@@ -9,10 +8,9 @@ export { StaticSeekError } from "@src/frontend/base";
 export type {
     StaticSeekIndex,
     StaticSeekIndexObject,
-    SearchEnv,
+    IndexOpt,
     FieldName,
     Path,
-    FieldNameMap,
     Reference,
     SearchResult,
     SearchFn,